Login and registration page added with UI, they have validation and user authentication. Started working on the UI of the chatting page
The login and registration pages are finished and redirect users into the application. Proper error messages are added which are displayed in url. Same usernames are not allowed, and when registering the user's password and salt is hashed and stored in the database. I am still working on the UI of the chatting page, its mostly done but I am having trouble connecting it to phase 1 and 2